EEF.profile {boot}	R Documentation

Empirical Likelihoods

Description

Construct the empirical log likelihood or empirical exponential family log likelihood for a mean.

Usage

EEF.profile(y, tmin = min(y) + 0.1, tmax = max(y) - 0.1, n.t = 25, 
            u = function(y, t) y - t)
EL.profile(y, tmin  =  min(y) + 0.1, tmax = max(y) - 0.1, n.t  =  25, 
           u = function(y, t) y - t)

Arguments

y
A vector or matrix of data

tmin
The minimum value of the range over which the likelihood should be computed. This must be larger than min(y).

tmax
The maximum value of the range over which the likelihood should be computed. This must be smaller than max(y).

n.t
The number of points between tmin and tmax at which the value of the log-likelihood should be computed.

u
A function of the data and the parameter.

Details

These functions calculate the log likelihood for a mean using either an empirical likelihood or an empirical exponential family likelihood. They are supplied as part of the package boot for demonstration purposes with the practicals in chapter 10 of Davison and Hinkley (1997). The functions are not intended for general use and are not supported as part of the bootpackage. For more general and more robust code to calculate empirical likelihoods see Professor A. B. Owen's empirical likelihood home page at the URL http://statistics.stanford.edu/~owen/empirical/.

Value

A matrix with n.t rows. The first column contains the values of the parameter used. The second column of the output of EL.profile contains the values of the empirical log likelihood. The second and third columns of the output of EEF.profile contain two versions of the empirical exponential family log-likelihood. The final column of the output matrix contains the values of the Lagrange multiplier used in the optimization procedure.

Author(s)

Angelo J. Canty

References

Davison, A. C. and Hinkley, D. V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.

[Package boot version 1.3-17 Index]