Bootstrap Resampling Arrays

This function takes a bootstrap object calculated by one of the functions boot, censboot, or tilt.boot and returns the frequency (or index) array for the bootstrap resamples.

boot.array(boot.out, indices)

Arguments

boot.out
An object of class "boot" returned by one of the generation functions for such an object.

indices
A logical argument which specifies whether to return the frequency array or the raw index array. The default is indices=FALSE unless boot.out was created by tsboot in which case the default is indices=TRUE.

Value

A matrix with boot.out$R rows and n columns where n is the number of observations in boot.out$data. If indices is FALSE then this will give the frequency of each of the original observations in each bootstrap resample. If indices is TRUE it will give the indices of the bootstrap resamples in the order in which they would have been passed to the statistic.

Side Effects

This function temporarily resets .Random.seed to the value in boot.out$seed and then returns it to its original value at the end of the function.

Details

The process by which the original index array was generated is repeated with the same value of .Random.seed. If the frequency array is required then freq.array is called to convert the index array to a frequency array.

A resampling array can only be returned when such a concept makes sense. In particular it cannot be found for any parametric or model-based resampling schemes. Hence for objects generated by censboot the only resampling scheme for which such an array can be found is ordinary case resampling. Similarly if boot.out$sim is "parametric" in the case of boot or "model" in the case of tsboot the array cannot be found. Note also that for post-blackened bootstraps from tsboot the indices found will relate to those prior to any post-blackening and so will not be useful.

Frequency arrays are used in many post-bootstrap calculations such as the jackknife-after-bootstrap and finding importance sampling weights. They are also used to find empirical influence values through the regression method.

See also

Examples

#  A frequency array for a nonparametric bootstrap
city.boot <- boot(city, corr, R = 40, stype = "w")
boot.array(city.boot)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    2    1    0    0    2    0    2    1     2
#>  [2,]    1    2    1    0    2    0    1    2    1     0
#>  [3,]    1    0    1    1    1    1    2    2    1     0
#>  [4,]    1    1    0    1    1    0    1    0    1     4
#>  [5,]    1    2    0    1    1    2    0    0    2     1
#>  [6,]    1    1    2    3    0    0    0    2    1     0
#>  [7,]    3    0    0    0    0    0    1    0    3     3
#>  [8,]    1    0    2    0    1    2    0    2    1     1
#>  [9,]    1    2    0    1    0    1    1    0    1     3
#> [10,]    1    1    1    1    1    2    1    1    0     1
#> [11,]    2    1    1    1    1    0    2    0    1     1
#> [12,]    1    0    1    0    3    2    2    1    0     0
#> [13,]    1    0    0    1    1    1    0    1    4     1
#> [14,]    0    0    1    1    1    3    0    3    1     0
#> [15,]    2    0    1    0    2    2    0    0    1     2
#> [16,]    0    0    2    0    2    2    0    0    2     2
#> [17,]    0    0    1    3    0    0    1    2    3     0
#> [18,]    0    1    1    0    0    2    3    2    1     0
#> [19,]    0    2    1    0    1    2    0    1    1     2
#> [20,]    0    0    1    2    5    1    0    1    0     0
#> [21,]    0    0    0    1    0    2    0    2    2     3
#> [22,]    0    0    1    1    1    4    0    0    2     1
#> [23,]    2    1    1    2    1    0    1    0    2     0
#> [24,]    0    2    0    3    0    0    4    0    0     1
#> [25,]    0    0    0    1    0    1    3    1    3     1
#> [26,]    2    0    2    1    1    0    0    1    1     2
#> [27,]    1    2    1    0    1    1    3    1    0     0
#> [28,]    3    3    0    0    0    2    0    0    2     0
#> [29,]    0    1    0    2    0    0    3    1    1     2
#> [30,]    3    0    0    0    3    1    0    1    0     2
#> [31,]    1    2    1    2    0    0    0    1    2     1
#> [32,]    3    0    1    1    2    0    1    1    0     1
#> [33,]    1    1    0    0    1    1    2    4    0     0
#> [34,]    0    2    2    0    0    2    0    2    1     1
#> [35,]    0    0    0    1    2    1    2    2    1     1
#> [36,]    0    0    1    1    0    1    2    2    2     1
#> [37,]    0    3    0    0    2    2    1    0    2     0
#> [38,]    2    2    1    1    1    1    0    1    1     0
#> [39,]    2    2    0    0    0    0    1    2    2     1
#> [40,]    1    0    0    0    2    2    2    1    1     1

perm.cor <- function(d,i) cor(d$x,d$u[i])
city.perm <- boot(city, perm.cor, R = 40, sim = "permutation")
boot.array(city.perm, indices = TRUE)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    2    9   10    8    7    5    4    3    1     6
#>  [2,]    4    8    7    9    6   10    1    5    2     3
#>  [3,]    7    8    9    1    6    5    3    2    4    10
#>  [4,]    6    8    5    4    2   10    3    1    7     9
#>  [5,]    5    4    2    8    1    3   10    7    6     9
#>  [6,]    6    2    8    1    3   10    9    7    5     4
#>  [7,]    8    9    6   10    1    7    5    3    4     2
#>  [8,]    5    2    4    6    1   10    8    9    3     7
#>  [9,]    2    4    6    5    8    7   10    1    3     9
#> [10,]   10    5    2    1    7    3    9    8    4     6
#> [11,]   10    9    6    5    1    7    4    2    3     8
#> [12,]    6    5    1   10    7    9    4    8    3     2
#> [13,]    2    7    4    3    5    6    9    1   10     8
#> [14,]   10    2    6    4    8    5    7    3    9     1
#> [15,]    8    2    9    3    4   10    1    6    5     7
#> [16,]    2    6    7    4    1    5    9    3    8    10
#> [17,]    4    8    6    3    9    5   10    7    1     2
#> [18,]    8    4    3    6   10    1    2    5    9     7
#> [19,]    9    5    2   10    3    4    8    7    1     6
#> [20,]    3    8    2    4   10    1    7    5    6     9
#> [21,]    7    6    8    5    9    1   10    2    3     4
#> [22,]    4    6   10    5    7    2    9    1    3     8
#> [23,]    2    3    7    4    8   10    1    5    9     6
#> [24,]    6    8    5   10    3    1    4    2    9     7
#> [25,]    9    2    5    1    6    4    3   10    8     7
#> [26,]    5    7    6    8    1    3    2    4   10     9
#> [27,]   10    8    2    6    3    9    5    7    4     1
#> [28,]   10    2    4    6    1    8    5    7    9     3
#> [29,]    8    1    6    5    7    2    4   10    9     3
#> [30,]    7    8    3    5    4    6    1    2    9    10
#> [31,]    1   10    8    3    4    6    9    7    2     5
#> [32,]    6   10    3    5    9    4    1    8    7     2
#> [33,]    2    4    7    5    3    6    9    8    1    10
#> [34,]    8    5    1    9    4   10    2    6    7     3
#> [35,]    4    7   10    1    5    3    6    9    8     2
#> [36,]    1    6    2    3    9    5    8    7    4    10
#> [37,]    1    8    9    5    2   10    7    6    3     4
#> [38,]    4    9    1    2    6    7    3    8   10     5
#> [39,]    5   10    8    6    2    9    3    4    1     7
#> [40,]    2    7    6    4    3    1    8    9    5    10