glm.diag {boot}	R Documentation

Generalized Linear Model Diagnostics

Description

Calculates jackknife deviance residuals, standardized deviance residuals, standardized Pearson residuals, approximate Cook statistic, leverage and estimated dispersion.

Usage


glm.diag(glmfit)

Arguments

glmfit
glmfit is a glm.object - the result of a call to glm()

Value

Returns a list with the following components

res
The vector of jackknife deviance residuals.

rd
The vector of standardized deviance residuals.

rp
The vector of standardized Pearson residuals.

cook
The vector of approximate Cook statistics.

h
The vector of leverages of the observations.

sd
The value used to standardize the residuals. This is the estimate of residual standard deviation in the Gaussian family and is the square root of the estimated shape parameter in the Gamma family. In all other cases it is 1.

Note

See the help for glm.diag.plots for an example of the use of glm.diag.

References

Davison, A.C. and Snell, E.J. (1991) Residuals and diagnostics. In Statistical Theory and Modelling: In Honour of Sir David Cox. D.V. Hinkley, N. Reid and E.J. Snell (editors), 83–106. Chapman and Hall.

See Also

glm, glm.diag.plots, summary.glm

[Package boot version 1.3-20 Index]