inv.logit {boot}	R Documentation

Inverse Logit Function

Description

Given a numeric object return the inverse logit of the values.

Usage

inv.logit(x)

Arguments

x
A numeric object. Missing values (NAs) are allowed.

Details

The inverse logit is defined by exp(x)/(1+exp(x)). Values in x of -Inf or Inf return logits of 0 or 1 respectively. Any NAs in the input will also be NAs in the output.

Value

An object of the same type as x containing the inverse logits of the input values.

See Also

logit, plogis for which this is a wrapper.

[Package boot version 1.3-17 Index]