print.bootci {boot}	R Documentation

Print Bootstrap Confidence Intervals

Description

This is a method for the function print() to print objects of the class "bootci".

Usage


## S3 method for class 'bootci'
print(x, hinv = NULL, ...)

Arguments

x
The output from a call to boot.ci.

hinv
A transformation to be made to the interval end-points before they are printed.

...
further arguments passed to or from other methods.

Details

This function prints out the results from boot.ci in a "nice" format. It also notes whether the scale of the intervals is the original scale of the input to boot.ci or a different scale and whether the calculations were done on a transformed scale. It also looks at the order statistics that were used in calculating the intervals. If the smallest or largest values were used then it prints a message

Warning : Intervals used Extreme Quantiles

Such intervals should be considered very unstable and not relied upon for inferences. Even if the extreme values are not used, it is possible that the intervals are unstable if they used quantiles close to the extreme values. The function alerts the user to intervals which use the upper or lower 10 order statistics with the message

Some intervals may be unstable

Value

The object ci.out is returned invisibly.

See Also

boot.ci

[Package boot version 1.3-20 Index]