print.saddle.distn {boot}	R Documentation

Print Quantiles of Saddlepoint Approximations

Description

This is a method for the function print() to print objects of class "saddle.distn".

Usage


## S3 method for class 'saddle.distn'
print(x, ...)

Arguments

x
An object of class "saddle.distn" created by a call to saddle.distn.

...
further arguments passed to or from other methods.

Details

The quantiles of the saddlepoint approximation to the distribution are printed along with the original call and some other useful information.

Value

The input is returned invisibly.

See Also

lines.saddle.distn, saddle.distn

[Package boot version 1.3-20 Index]