print.simplex {boot}	R Documentation

Print Solution to Linear Programming Problem

Description

This is a method for the function print() to print objects of class "simplex".

Usage


## S3 method for class 'simplex'
print(x, ...)

Arguments

x
An object of class "simplex" created by calling the function simplex to solve a linear programming problem.

...
further arguments passed to or from other methods.

Details

The coefficients of the objective function are printed. If a solution to the linear programming problem was found then the solution and the optimal value of the objective function are printed. If a feasible solution was found but the maximum number of iterations was exceeded then the last feasible solution and the objective function value at that point are printed. If no feasible solution could be found then a message stating that is printed.

Value

x is returned silently.

See Also

simplex

[Package boot version 1.3-20 Index]