saddle.distn.object {boot}	R Documentation

Saddlepoint Distribution Approximation Objects

Description

Class of objects that result from calculating saddlepoint distribution approximations by a call to saddle.distn.

Generation

This class of objects is returned from calls to the function saddle.distn.

Methods

The class "saddle.distn" has methods for the functions lines and print.

Structure

Objects of class "saddle.distn" are implemented as a list with the following components.

quantiles

A matrix with 2 columns. The first column contains the probabilities alpha and the second column contains the estimated quantiles of the distribution at those probabilities derived from the spline.

points

A matrix of evaluations of the saddlepoint approximation. The first column contains the values of t which were used, the second and third contain the density and cdf approximations at those points and the rest of the columns contain the solutions to the saddlepoint equations. When type is "simp", there is only one of those. When type is "cond" there are 2*d-1 where d is the number of columns in A or the output of A(t,...{}). The first d of these correspond to the numerator and the remainder correspond to the denominator.

distn

An object of class smooth.spline. This corresponds to the spline fitted to the saddlepoint cdf approximations in points in order to approximate the entire distribution. For the structure of the object see smooth.spline.

call

The original call to saddle.distn which generated the object.

LR

A logical variable indicating whether the Lugananni-Rice approximations were used.

See Also

lines.saddle.distn, saddle.distn, print.saddle.distn

[Package boot version 1.3-20 Index]