urine {boot}	R Documentation

Urine Analysis Data

Description

The urine data frame has 79 rows and 7 columns.

79 urine specimens were analyzed in an effort to determine if certain physical characteristics of the urine might be related to the formation of calcium oxalate crystals.

Usage

urine

Format

This data frame contains the following columns:

r

Indicator of the presence of calcium oxalate crystals.

gravity

The specific gravity of the urine.

ph

The pH reading of the urine.

osmo

The osmolarity of the urine. Osmolarity is proportional to the concentration of molecules in solution.

cond

The conductivity of the urine. Conductivity is proportional to the concentration of charged ions in solution.

urea

The urea concentration in millimoles per litre.

calc

The calcium concentration in millimoles per litre.

Source

The data were obtained from

Andrews, D.F. and Herzberg, A.M. (1985) Data: A Collection of Problems from Many Fields for the Student and Research Worker. Springer-Verlag.

References

Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.

[Package boot version 1.2-38 Index]