boot-practicals {boot}	R Documentation

Functions for Bootstrap Practicals

Description

Functions for use with the practicals in Davison and Hinkley (1997)

Usage

nested.corr(data, w, t0, M)
lik.CI(like, lim)

Details

nested.corr is meant for use with the double bootstrap in practical 5.5 of Davison and Hinkley (1997).

lik.CI is meant for use with practicals 10.1 and 10.2 of Davison and Hinkley (1997).

Author(s)

Angelo J. Canty

References

Davison, A. C. and Hinkley, D. V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.

[Package boot version 1.2-38 Index]