cd4.nested {boot}	R Documentation

Nested Bootstrap of cd4 data

Description

This is an example of a nested bootstrap for the correlation coefficient of the cd4 data frame. It is used in a practical in Chapter 5 of Davison and Hinkley (1997).

References

Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.

See Also

cd4

[Package boot version 1.3-28 Index]