cum3 {boot}	R Documentation

Calculate Third Order Cumulants

Description

Calculates an estimate of the third cumulant, or skewness, of a vector. Also, if more than one vector is specified, a product-moment of order 3 is estimated.

Usage


cum3(a, b = a, c = a, unbiased = TRUE)

Arguments

a
A vector of observations.

b
Another vector of observations, if not supplied it is set to the value of a. If supplied then it must be the same length as a.

c
Another vector of observations, if not supplied it is set to the value of a. If supplied then it must be the same length as a.

unbiased
A logical value indicating whether the unbiased estimator should be used.

Details

The unbiased estimator uses a multiplier of n/((n-1)*(n-2)) where n is the sample size, if unbiased is FALSE then a multiplier of 1/n is used. This is multiplied by sum((a-mean(a))*(b-mean(b))*(c-mean(c))) to give the required estimate.

Value

The required estimate.

[Package boot version 1.3-20 Index]