k3.linear {boot}	R Documentation

Linear Skewness Estimate

Description

Estimates the skewness of a statistic from its empirical influence values.

Usage

k3.linear(L, strata=NULL)

Arguments

L 	 Vector of the empirical influence values of a statistic. These will usually be calculated by a call to empinf.

strata 	 A numeric vector or factor specifying which observations (and hence which components of L) come from which strata.

Value

The skewness estimate calculated from L.

References

Davison, A. C. and Hinkley, D. V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.

See Also

empinf, linear.approx, var.linear

Examples

#  To estimate the skewness of the ratio of means for the city data.
ratio <- function(d,w)
     sum(d$x * w)/sum(d$u * w)
k3.linear(empinf(data=city,statistic=ratio))

[Package boot version 1.2-38 Index]