logit {boot}	R Documentation

Logit of Proportions

Description

This function calculates the logit of proportions.

Usage

logit(p)

Arguments

p 	 A numeric Splus object, all of whose values are in the range [0,1]. Missing values (NAs) are allowed.

Details

If any elements of p are outside the unit interval then an error message is generated. Values of p equal to 0 or 1 (to within machine precision) will return -Inf or Inf respectively. Any NAs in the input will also be NAs in the output.

Value

A numeric object of the same type as p containing the logits of the input values.

See Also

inv.logit

[Package boot version 1.2-38 Index]