print.boot {boot}	R Documentation

Print a Summary of a Bootstrap Object

Description

This is a method for the function print() for objects of the class "boot" created by a call to boot, censboot, tilt.boot or tsboot.

Usage

## S3 method for class 'boot'
print(x, digits = getOption("digits"), 
      index = 1:ncol(boot.out$t), ...)

Arguments

x
A bootstrap output object of class "boot" generated by one of the bootstrap functions.

digits
The number of digits to be printed in the summary statistics.

index
Indices indicating for which elements of the bootstrap output summary statistics are required.

...
further arguments passed to or from other methods.

Details

For each statistic calculated in the bootstrap the original value and the bootstrap estimates of its bias and standard error are printed. If boot.out$t0 is missing (such as when it was created by a call to tsboot with orig.t = FALSE) the bootstrap mean and standard error are printed. If resampling was done using importance resampling weights, then the bootstrap estimates are reweighted as if uniform resampling had been done. The ratio importance sampling estimates are used and if there were a number of distributions then defensive mixture distributions are used. In this case an extra column with the mean of the observed bootstrap statistics is also printed.

Value

The bootstrap object is returned invisibly.

See Also

boot, censboot, imp.moments, plot.boot, tilt.boot, tsboot

[Package boot version 1.3-28 Index]