simplex.object {boot}	R Documentation

Linear Programming Solution Objects

Description

Class of objects that result from solving a linear programming problem using simplex.

Generation

This class of objects is returned from calls to the function simplex.

Methods

The class "saddle.distn" has a method for the function print.

Structure

Objects of class "simplex" are implemented as a list with the following components.

soln

The values of x which optimize the objective function under the specified constraints provided those constraints are jointly feasible.

solved

This indicates whether the problem was solved. A value of -1 indicates that no feasible solution could be found. A value of 0 that the maximum number of iterations was reached without termination of the second stage. This may indicate an unbounded function or simply that more iterations are needed. A value of 1 indicates that an optimal solution has been found.

value

The value of the objective function at soln.

val.aux

This is NULL if a feasible solution is found. Otherwise it is a positive value giving the value of the auxiliary objective function when it was minimized.

obj

The original coefficients of the objective function.

a

The objective function coefficients re-expressed such that the basic variables have coefficient zero.

a.aux

This is NULL if a feasible solution is found. Otherwise it is the re-expressed auxiliary objective function at the termination of the first phase of the simplex method.

A

The final constraint matrix which is expressed in terms of the non-basic variables. If a feasible solution is found then this will have dimensions m1+m2+m3 by n+m1+m2, where the final m1+m2 columns correspond to slack and surplus variables. If no feasible solution is found there will be an additional m1+m2+m3 columns for the artificial variables introduced to solve the first phase of the problem.

basic

The indices of the basic (non-zero) variables in the solution. Indices between n+1 and n+m1 correspond to slack variables, those between n+m1+1 and n+m2 correspond to surplus variables and those greater than n+m2 are artificial variables. Indices greater than n+m2 should occur only if solved is -1 as the artificial variables are discarded in the second stage of the simplex method.

slack

The final values of the m1 slack variables which arise when the "<=" constraints are re-expressed as the equalities A1%*%x + slack = b1.

surplus

The final values of the m2 surplus variables which arise when the "<=" constraints are re-expressed as the equalities A2%*%x - surplus = b2.

artificial

This is NULL if a feasible solution can be found. If no solution can be found then this contains the values of the m1+m2+m3 artificial variables which minimize their sum subject to the original constraints. A feasible solution exists only if all of the artificial variables can be made 0 simultaneously.

See Also

print.simplex, simplex

[Package boot version 1.2-38 Index]