var.linear {boot}	R Documentation

Linear Variance Estimate

Description

Estimates the variance of a statistic from its empirical influence values.

Usage

var.linear(L, strata=NULL)

Arguments

L 	 Vector of the empirical influence values of a statistic. These will usually be calculated by a call to empinf.

strata 	 A numeric vector or factor specifying which observations (and hence empirical influence values) come from which strata.

Value

The variance estimate calculated from L.

References

Davison, A. C. and Hinkley, D. V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.

See Also

empinf, linear.approx, k3.linear

Examples

#  To estimate the variance of the ratio of means for the city data.
ratio <- function(d,w)
     sum(d$x * w)/sum(d$u * w)
var.linear(empinf(data=city,statistic=ratio))

[Package boot version 1.2-38 Index]